[SIDE-479] Extract common members for GRPC Host Client
